# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TTDMI - Tourism Trust & Digital Maturity Index API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Operator {
  id: number;
  legal_name: string;
  commercial_name: string;
  /** @nullable */
  ruc?: string | null;
  region: string;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  district?: string | null;
  operator_type: string;
  /** @nullable */
  niche?: string | null;
  languages?: string[];
  verified: boolean;
  ttdmi_score: number;
  level: string;
  level_label?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  rank_nacional?: number | null;
  created_at: string;
}

export type TtdmiScoreScoreDetails = { [key: string]: unknown };

export interface TtdmiScore {
  operator_id: number;
  total_score: number;
  reviews_score: number;
  social_score: number;
  website_score: number;
  seo_score: number;
  engagement_score: number;
  formalidad_score: number;
  conversion_score: number;
  freshness_score: number;
  network_score: number;
  validation_factor: number;
  level: string;
  level_label?: string;
  score_details?: TtdmiScoreScoreDetails;
  updated_at: string;
}

export interface SocialProfile {
  id: number;
  platform: string;
  /** @nullable */
  handle?: string | null;
  followers: number;
  engagement_rate: number;
  verified: boolean;
  /** @nullable */
  last_post_date?: string | null;
  /** @nullable */
  posts_per_week?: number | null;
}

export interface ReviewSource {
  source: string;
  count: number;
  avg_rating: number;
}

export interface ReviewsSummary {
  average_rating: number;
  total_reviews: number;
  authenticity_score: number;
  sources?: ReviewSource[];
  sentiment_score?: number;
  recency_score?: number;
}

export interface BookingCapabilities {
  booking_engine: boolean;
  online_payment: boolean;
  whatsapp_cta: boolean;
  chatbot: boolean;
  lead_form: boolean;
  /** @nullable */
  response_time_hours?: number | null;
  ota_presence?: string[];
}

export interface Certification {
  id: number;
  name: string;
  issuer: string;
  verified: boolean;
  /** @nullable */
  expiry_date?: string | null;
}

export interface OperatorDetail {
  id: number;
  legal_name: string;
  commercial_name: string;
  /** @nullable */
  ruc?: string | null;
  region: string;
  /** @nullable */
  province?: string | null;
  /** @nullable */
  district?: string | null;
  operator_type: string;
  /** @nullable */
  niche?: string | null;
  languages?: string[];
  verified: boolean;
  ttdmi_score: number;
  level: string;
  level_label?: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  rank_nacional?: number | null;
  created_at: string;
  score_breakdown: TtdmiScore;
  social_profiles?: SocialProfile[];
  reviews_summary?: ReviewsSummary;
  booking_capabilities?: BookingCapabilities;
  certifications?: Certification[];
}

export interface OperatorListResponse {
  data: Operator[];
  total: number;
  page: number;
  limit: number;
}

export interface OperatorInput {
  legal_name: string;
  commercial_name: string;
  ruc?: string;
  region: string;
  province?: string;
  district?: string;
  operator_type: string;
  niche?: string;
  languages?: string[];
  website?: string;
  description?: string;
  phone?: string;
  email?: string;
}

export interface OperatorUpdate {
  legal_name?: string;
  commercial_name?: string;
  ruc?: string;
  region?: string;
  province?: string;
  district?: string;
  operator_type?: string;
  niche?: string;
  languages?: string[];
  website?: string;
  description?: string;
  phone?: string;
  email?: string;
  verified?: boolean;
}

export interface RankingEntry {
  rank: number;
  operator_id: number;
  commercial_name: string;
  region: string;
  operator_type: string;
  /** @nullable */
  niche?: string | null;
  ttdmi_score: number;
  level: string;
  level_label?: string;
  verified: boolean;
  /** @nullable */
  logo_url?: string | null;
  /** @nullable */
  rank_change?: number | null;
}

export interface RankingResponse {
  type: string;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  niche?: string | null;
  data: RankingEntry[];
  generated_at: string;
}

export interface AnalyticsSummary {
  total_operators: number;
  verified_operators: number;
  avg_ttdmi_score: number;
  elite_count: number;
  premium_count: number;
  advanced_count: number;
  growing_count: number;
  emerging_count: number;
  risk_count: number;
  regions_covered: number;
  top_region: string;
  top_niche: string;
  digital_gap_operators?: number;
  hidden_gems_count?: number;
}

export interface ScoreDistributionItem {
  level: string;
  label: string;
  count: number;
  avg_score: number;
  color: string;
}

export interface RegionStat {
  region: string;
  count: number;
  avg_score: number;
  verified_count?: number;
  top_operator: string;
  elite_count?: number;
}

export interface NicheStat {
  niche: string;
  count: number;
  avg_score: number;
  /** @nullable */
  top_operator?: string | null;
  /** @nullable */
  growth_trend?: string | null;
}

export interface HiddenGem {
  operator_id: number;
  commercial_name: string;
  region: string;
  operator_type: string;
  ttdmi_score: number;
  potential_score: number;
  gap_score: number;
  insight: string;
  /** @nullable */
  logo_url?: string | null;
}

export interface DigitalGapOperator {
  operator_id: number;
  commercial_name: string;
  region: string;
  operator_type: string;
  reviews_score: number;
  digital_score: number;
  gap: number;
  recommendation: string;
  /** @nullable */
  logo_url?: string | null;
}

export interface ScoreTrendItem {
  month: string;
  avg_score: number;
  total_operators: number;
  elite_count?: number;
  premium_count?: number;
}

export type ListOperatorsParams = {
search?: string;
region?: string;
operator_type?: string;
niche?: string;
min_score?: number;
max_score?: number;
verified?: boolean;
level?: string;
page?: number;
limit?: number;
sort_by?: string;
};

export type GetRankingsParams = {
/**
 * Ranking type
 */
type?: GetRankingsType;
region?: string;
niche?: string;
limit?: number;
};

export type GetRankingsType = typeof GetRankingsType[keyof typeof GetRankingsType];


export const GetRankingsType = {
  nacional: 'nacional',
  regional: 'regional',
  provincial: 'provincial',
  niche: 'niche',
  receptivo: 'receptivo',
  conversion: 'conversion',
  trust: 'trust',
  seo: 'seo',
  freshness: 'freshness',
  hidden_gems: 'hidden_gems',
} as const;

export type GetHiddenGemsParams = {
limit?: number;
};

export type GetDigitalGapsParams = {
limit?: number;
};