# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TTDMI - Tourism Trust & Digital Maturity Index API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/hidden-gems:
    get:
      operationId: getHiddenGems
      tags: [analytics]
      summary: Detect undervalued high-potential operators
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Hidden gem operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HiddenGem"
  /analytics/digital-gaps:
    get:
      operationId: getDigitalGaps
      tags: [analytics]
      summary: Operators with strong offline reputation but poor digitalization
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Digital gap operators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DigitalGapOperator"
...
components:
  schemas:
...
    HiddenGem:
      type: object
      required: [operator_id, commercial_name, region, operator_type, ttdmi_score, potential_score, gap_score, insight]
      properties:
        operator_id:
          type: integer
        commercial_name:
          type: string
        region:
          type: string
        operator_type:
          type: string
        ttdmi_score:
          type: number
        potential_score:
          type: number
        gap_score:
          type: number
        insight:
          type: string
        logo_url:
          type: ["string", "null"]
    DigitalGapOperator:
      type: object
      required: [operator_id, commercial_name, region, operator_type, reviews_score, digital_score, gap, recommendation]
      properties:
        operator_id:
          type: integer
        commercial_name:
          type: string
        region:
          type: string
        operator_type:
          type: string
        reviews_score:
          type: number
        digital_score:
          type: number
        gap:
          type: number
        recommendation:
          type: string
        logo_url:
          type: ["string", "null"]
...